refactored the code a bit in preparation for the import functions
